[YDB#1071]: Reduce the number of test pipelines and speed up overall pipeline by picking tests at random
- `docker/pipeline-test-ydb.csh` is a new script which was commited by
mistake in the previous commit e2248a0, but not used by it.
- Runs a pipline for each commit of YDB repo so that a random set of
  tests is run.
- `registry.gitlab.com/yottadb/db/ydbtest` now takes a new argument
  `-pipelineydb` for running the new script.

Sample pipeline:
https://gitlab.com/shabiel/YDB/-/jobs/8209443822 (6 minutes)

Future work: the 6 minute runtime should be able to be cut down to 2
minutes or so with extra work to run the script on a pre-downloaded
docker image. Right now, the docker image needs to be downloaded for
every run. This is tracked at
https://gitlab.com/YottaDB/DB/YDB/-/issues/1105.