New pre-commit feature: indentation check - warning for lines starting with space
This MR contains a new feature for _pre-commit_ script: indentation check.

It warns for committed files which contains lines starting with space instead of tab. If the file contains 10 lines or less of such ones, it prints the lines with line numbers, otherwise prints only filenames and number of occurrences.

The script works with whitelist; contains a list of filename extensions which should be checked:
```
set check_ext_list = "m csh sh c"
```

Example output:
```
Files with wrong indentation
  pre-commit (1 line found):
    > 149:                if ($file == $check_file) then
  v70004/inref/gtmde201295.m (4 lines found):
    > 15:    ; important: don't perform any database or print operation here
    > 19:        set beg=$zut
    > 21:        set end=$zut
    > 22:    ;
  v70004/u_inref/honor_socket_connect_timeout-gtmde201295.csh (28 lines found)
```