Fixed `tst: Undefined variable.` error in com/lasttest.csh with `gtmtest.csh -bg`
Background
----------

Previously, `lasttest.csh` would issue a `tst: Undefined variable.` error when running `gtmtest.csh` in the background and immediately sourcing `lasttest.csh`.

Specifically, this error appears when:
1. `-fg` is omitted from `$gtmtest_args` in `.cshrc`.
2. a `gtmtest` alias such as the following is run: `tsync; tcsh -fc "$gtm_test_com_individual/gtmtest.csh $gtmtest_args !*:q"; source $gtm_test_com_individual/lasttest.csh`

The issue is caused by `lasttest.csh` attempting to access the directory containing the last test's results before that directory is created by `gtmtest.csh`. Specifically, line 24 below checks for the existence of the last test directory, and breaks the loop if it isn't there:

```sh
17 setenv last_tst_dir `ls -dt $tst_dir/tst_* | head -n1`
18 if ( -e $last_tst_dir/submitted_tests ) then
19         # If test was run with runtest or gtmtest, there is a submitted_tests file containing
20         # a list of tests that were run, so extract name of last test run.
21         # Otherwise (if run by runsubtest), retain $tst from what was previously set by settest.
22         foreach tst (`awk '{print $2}' $last_tst_dir/submitted_tests`)
23                 # Find the last listed test that was actually run
24                 if ! (-e $last_tst_dir/${tst}_0) break
25                 setenv tst $tst
26         end
27         unset tst   # unset the internal variable to expose the last env variable we set above
28 endif
29
30 if ( "$tst" == "" ) then
31         echo "\$last_tst_dir/submitted_tests should contain a valid list of tests that were run. Did they run?"
32         ln -fns $last_tst_dir "$r"   # if we can't find a specific test dir, at least point to its parent
33         exit 1
34 endif
```

Under the above mentioned conditions, this code block is executed before `submittest.csh` (run by `gtmtest.csh`) is finished running in the background and before the last test directory is created. This means that line 25 won't execute, leaving the `$tst` environment variable undefined, leading in turn to the `tst: Undefined variable.` error.

Fix
---

The fix is to wait for the creation the last run test directory before attempting to set the `$tst` environment variable in `lasttest.csh`. This is now done with a loop that checks for that directory's existence and waits if it doesn't exist, up to a maximum of 60 seconds. Once the directory is found, code execution proceeds and the `$tst` environment variable is set to the name of the last test, as previously.

Additionally, the `$tst` *local* variable was renamed to `$submitted_tst` to avoid ambiguity with the `$tst` environment variable and prevent reliance on the shell implicitly determining whether to access the local or environment variable `$tst`. This also allowed the removal of an `unset tst` line, since it is no longer necessary to explicitly unset the local `$tst` variable to force the implicit usage of the `$tst` environment variable.

Note also that this fix only applies in the simple case of running a single test without replication due to the check for `${tst}_0` on line 24 above. With this fix, the original `tst: Undefined variable.` error will still occur if either of the following conditions are met:
+ Submitting a test with `-replic`, which would only create `${tst}_1`
+ Submitting a test with `-num_runs 10` etc., which would only create `${tst}_0_1` etc.