Added firefox to list of exclusions in com/lsof to prevent test slowdown
Background
----------

Significant test slowdowns of were observed when the Firefox web browser was run concurrently with the YottaDB test system. This was due to the `lsof` call in `com/lsof` outputting thousands of lines due to open files held by the `firefox` process during test execution.

Fix
---

See commit message of 3cb251e7 for similar issue and how it was fixed for `chrome` browser.

The fix is to add a `-c "^firefox` line to the `lsof` call in `com/lsof` to exclude files opened by Firefox from `lsof` output.