[#659] Fixed v71001/randstr_rangearg-gtmde500856 'if: Expression Syntax.' failure
A test failure in `v71001/randstr_rangearg-gtmde500856` was reported on an in-house system with the following outref diff:

```diff
16,17c16
< PASS: Took less than 150000000 instructions
<
---
> if: Expression Syntax.
21,22c20,21
< PASS: Took less than 150000000 instructions
<
---
> if: Expression Syntax.
> continue: end not found.
```

This was due to the `$instructions` array containing different values depending on whether the test was run in `UTF-8` mode. In `UTF-8` mode, an `INVDLRCVAL` error was issued causing the prompt prefix to be directed to `stdout`. In that case, the `perf` output stored in `$instructions` omitted the prompt, such that `$instructions[1]` contained the number of instructions instead of the prompt prefix.

The existing check on line 59 did not properly account for this and assumed that if `$instructions[1]` was not set to `GTM>`, then no prompt was present. In that case, `$total_instructions` was set to `$instructions[1]` on the assumption that the latter contained the number of instructions.

However, it is possible that the prompt was `YDB>`. In that case, the check on line 59 would fail, but `$instructions[1]` would be set to `YDB>` instead of the number of instructions, causing the test failure.

This commit fixes this issue by redirecting both `stderr` and `stdout` of the `perf` output to a file, then reading the relevant line into `$instructions`. This allows the output to be handled uniformly, regardless of whether UTF-8 mode is set and an `INVDLRCVAL` error is issued.