[#787] Fix rsyslog related failures
1. Fix rsyslog detection of log message in date range

There is fuzziness logic for Systemd dates that is also apparently
needed for rsyslog as well. So it's now added.

This fixes mupjnl/C9C04001977 and encryption/eotf_basic.

2. Fix r122/tprestart so it runs with rsyslog

This test uses awk with space position numbers to get the information
from the syslog. SystemD dates come with spaces; rsyslog dates don't
have spaces, and that is what causes the difference.