[#787] Fix occassional failure in r136/ydb908 in Docker
Test failed with following diff:
```
r136 test starts...
FAIL from ydb908. Please check ydb908/ydb908.diff and
ydb908/errs_found.logx for errors in time sorted fashion
r136 test DONE.
00:00:00:02
\# Diff of ydb908 follows
28,40c28,31
< PASS
< # Test $ZGETJPI(1,UTIME) matches what exists in /proc/1/stat for <init> process (i.e. PID=1)
< PASS
< # Test $ZGETJPI(PID,CPUTIM) does not SIG-11 or assert fail or error for arbitrary list of PIDs found to exist in the system
< PASS
< # Test $ZGETJPI(PID,CSTIME) does not SIG-11 or assert fail or error for arbitrary list of PIDs found to exist in the system
< PASS
< # Test $ZGETJPI(PID,CUTIME) does not SIG-11 or assert fail or error for arbitrary list of PIDs found to exist in the system
< PASS
< # Test $ZGETJPI(PID,STIME) does not SIG-11 or assert fail or error for arbitrary list of PIDs found to exist in the system
< PASS
< # Test $ZGETJPI(PID,UTIME) does not SIG-11 or assert fail or error for arbitrary list of PIDs found to exist in the system
< PASS
---
> %YDB-E-DIVZERO, Attempt to divide by zero
> 		At M source location timeIsAlmostSame+5^ydb908

```

This test fails when STIME for PID 1 (test 3 in ydb908) is zero. That's
a valid outcome, but the test uses 0 as the denominator in a division in
function `$$timeIsAlmostSame^ydb908`.

```
test3 STIME 0 0
%YDB-E-DIVZERO, Attempt to divide by zero
              At M source location timeIsAlmostSame+5^ydb908

YDB>
```

In docker, the init process (PID 1) actually just started, whereas on a real
system it would have been running for a while. That's why STIME in
docker comes out with 0 or 1.

The solution is to change the function `$$timeIsAlmostSame` to allow
another sort of comparison that won't fail with a zero.

PS: other changes are just whitespace changes caught by a pre-commit
script.