[#787] Fix mupjnl/test_verbose for different timezones
Running this test in docker (which has an unset timezone, i.e.
defaulting to UTC) showed the following failure:

```
169c169
< ##DIFFERENT EXPRESSION##^.*%YDB-I-MUINFOUINT6,     Tp_resolve_time : 0 .0x00000000. ; \$H=47116,68400$## VS ##%YDB-I-MUINFOUINT6,     Tp_resolve_time : 0 [0x00000000] ; $H=47117,0##
---
> %YDB-I-MUINFOUINT6,     Tp_resolve_time : 0 [0x00000000] ; $H=47117,0
346c346
< ##DIFFERENT EXPRESSION##^.*%YDB-I-MUINFOUINT6,     Tp_resolve_time : 0 .0x00000000. ; \$H=47116,68400$## VS ##%YDB-I-MUINFOUINT6,     Tp_resolve_time : 0 [0x00000000] ; $H=47117,0##
---
> %YDB-I-MUINFOUINT6,     Tp_resolve_time : 0 [0x00000000] ; $H=47117,0
419c419
< ##DIFFERENT EXPRESSION##^.*%YDB-I-MUINFOUINT6,     Tp_resolve_time : 0 .0x00000000. ; \$H=47116,68400$## VS ##%YDB-I-MUINFOUINT6,     Tp_resolve_time : 0 [0x00000000] ; $H=47117,0##
---
> %YDB-I-MUINFOUINT6,     Tp_resolve_time : 0 [0x00000000] ; $H=47117,0
451c451
< ##DIFFERENT EXPRESSION##^.*%YDB-I-MUINFOUINT6,     Tp_resolve_time : 0 .0x00000000. ; \$H=47116,68400$## VS ##%YDB-I-MUINFOUINT6,     Tp_resolve_time : 0 [0x00000000] ; $H=47117,0##
---
> %YDB-I-MUINFOUINT6,     Tp_resolve_time : 0 [0x00000000] ; $H=47117,0
```

The cause of this failure is that YottaDB hardcodes a journal field to
be at the Unix Epoch, and this represents $H of 47116,68400--but only in
the Eastern Time Zone. The proper Unix Epoch is actually $H of 47117,0,
as the Unix Epoch is set at UTC. Running this test in different time
zones will actually produce different values. It's not desirable to have
the test output vary by timezone. This is solved by hardcoding the
timezone to UTC.