[#787] Fix v70000/ydb531_v6tov7 with rust
v70000/ydb531_v6tov7 is an imptp-using test. If imptp chooses rust, this
test fails, as it was not installed.

This commit installs rust and clang (which rust needs for compiling the
YDBRust wrapper).

PATH adjusted in a couple of spots as the Docker env doesn't propagate
to the gtmuser.

Otherwise, there are some cosmetic changes to the Dockerfile.