[#787] Fix sudo/shmhugetlb_syslog-gtmf221672 to run in docker
The test was docker ready, but the flag to test for to change behavior
is `$ydb_test_inside_docker`, not `$CI_PIPELINE_ID`, which is a pipeline
specific variable, and co-existent with the always-set `$ydb_test_inside_docker`.

Since this test does `setcap 'cap_ipc_lock+ep' $gtm_dist/dse`, outside
of the pipeline, which runs in priviliged mode, you need to pass
capability IPC_LOCK. README updated.