[#787] Fix replic pipeline failures due to regression in f2235674
Pipelines started to fail with this symptom:

```
TEST-E-REMOTE Remote directory could not be created. Exiting.
```

Putting some debugging info in `com/submit_test.csh` showed the
following:

```
/testarea2/gtmtest
TEST-E-REMOTE Remote directory could not be created. Exiting.
```

With f2235674, file `tstdir.csh` contains multiple test directories.
This caused the test system to now try using `/testarea2` for replic
tests, and since we don't do any set-up for this directory in the
pipeline (in particular `chmod 777`), we get permission errors.

Commit f2235674 said this:

> I decided not to update the pipeline to include failures in the
> artifacts from the new test areas from the "remote" hosts. It's
> doable if we need to in the future, but so far all pipeline tests
> failures have been reproducible locally, so there is never really
> a need to inspect the artifacts.

Because of this failure is essentially due to the fact that I didn't do
the point above, I went ahead and made that change as part of this
commit. Now the pipeline will have artifacts from all 3 testareas.

Here's an example of the previously failing pipeline succeeding:
https://gitlab.com/shabiel/YDBTest/-/jobs/9819196634.