[#787] Fix r130/ydb607 on Docker/pipeline
Test fails with following diff:

```
4a5
> VERSION-E-VERNOTEXIST : Directory /usr/library/RANDOMVER-E-CANNOTRUN/dbg does not exist. Exiting...
6c7
< Flush trigger                  960
---
> Flush trigger                          960
12a14,15
> /testarea1/tst_V999_R999_dbg_01_250515_031842_P8E/r130_0/ydb607/ver.log
> # Set version to: RANDOMVER-E-CANNOTRUN : Could not determine previous version matching the given criteria - -gte V63003 -lt V63007. Exiting...
```

The fix is for the instream.csh to check the available versions and
exclude the test if the versions are not present. This is a good fix not
just for docker/pipeline but any system where we try to run this test,
as the presence of previous versions is not guaranteed. Previous ways to
do this check have been removed, as this version is more robust.

To accomplish that, we needed to add an extra flag `-ck` to
`com/random_ver.csh` to tell it not to exit the whole script. See
updated help in file for detail.