[#828] Remove AARCH64 pipelines/images from YDBTest
Due to the fact that our machine for creating AARCH64 images is slow
(mostly due to IO speed), we decided to stop using it right now for
YDBTest as it slows down our development speed.

If and when we get a better machine, we can revert the commit and
re-enable the pipeline.

This reverts part of commit 440c1f5afc116d5a511cc0359568c2bbf07dcf2f,
which added that feature.

Running tests on the in-house AARCH64 systems on a regular basis
(weekly) is considered good enough for now.