[#787] Make TEST-E-NOTESTS error a warning and exit with 0
This change is done for a scenario where if TEST-W-APPLIC is issued, a
test is taken off the run list, and if it is the only test, then YDBTest
issues a TEST-E-NOTESTS message and exits with an error code of 14. This
is inconsistent with the general behavior of YDBTest, where if a test
cannot be run, it is simply skipped and the test run continues.

This scenario was seen with the following two tests on the pipeline:
- `dual_fail` without -replic
- `read_only` with -replic

Note that if there are multiple tests in the run list, and one of them
runs, then the whole run succeeds, even if TEST-W-APPLIC is issued.