[#807] Run specific YDB branch YDBTest pipeline
There will be two MRs for this issue:
- YDBTest pipeline pick a specific YDB branch (this MR)
- YDB pipeline pick a specific YDBTest branch

We use jq to get from the Gitlab API the YDB branches; if any match
the branch name of the YDBTest Merge Request, then the code for YDB
is fetched and built to be tested against.

The sudo test hack needed to be moved as it disturbs the build script
`build_and_install_yottadb.csh`, making it inoperative. This change was
only done for testing using `-rootshell` flag in `docker/run.csh`. We
build the new YottaDB version before the sudo hack in
`docker/pipeline-test-ydbtest.csh`, so it doesn't have to be moved.

Another small change is now we may need to get some git info from the
YDBTest repo, so an rsync script no longer excludes '.git' directory.