[YDB#1105] Sudo tests fail on YDB pipeline
Due to a regression in 0235bf2c5bf39baee1cc99874d4d03b954d544c6, sudo
tests requiring ydbinstall stopped working only inside the YDB pipeline
(YDBTest pipeline was still okay). See
https://gitlab.com/nars1/YDB/-/jobs/10224676723 for a sample failure.

That commit changed the how the pipeline runs, from running as a script
inside of a docker image after some set-up (which happened in
`docker/run.csh`) to running in a script directly invoked by the gitlab
runner. One of the steps that `docker/run.csh` does is to set-up a link
for an imaginary build directory for the use of sudo tests, as the sudo
tests are the only tests that use a file from the build directory. That
step is already in `docker/pipeline-test-ydbtest.csh`, but was forgotten
to be added to the modified file `docker/pipeline-test-ydb.csh`.