[#801] Fixed test level framework script failures in sudo/tmpyottadbperms-ydb1125 due to inadequate permissions
Previously, this test would pass at the subtest level, but fail at the test framework level due to a failure to clean up directories within the subtest due their being owned by root. This commit fixes this bug by changing the permissions on this directory before exiting the test, so that the test framework can clean it up.

Also, this commit disables this subtest in Docker, since it is currently guaranteed to fail. This is because of 2 reasons.
1. In docker /Distrib only contains the `ydbinstall` script but not the binaries to install (which usually is in case it is a full fledged staging area like is the case in in-house systems). This minimal setup in /Distrib is done to avoid the docker image space usage from bloating up considerably.
2. The `/Distrib/YottaDB/$1/$2/yottadb_r*/ydbinstall` invocation will go to gitlab and fetch the binaries if it does not find them in the current directory. This is okay as long as the test does not rely on the YDB master behavior and is fine with the latest production release (r2.02 at this point) behavior. This is the case for various subtests under the `sudo` test (e.g. ` sudo/u_inref/ydb910.sh`, `sudo/u_inref/ydb924.sh` etc.) and so those subtests continue to run fine even inside docker. But the `sudo/tmpyottadbperms-ydb1125` subtest relies on the YDB master behavior and so it will fail when the test runs with YDB r2.02 instead.

Additionally, a bug was fixed in a check on `$tmp_parent_dir` by revising the check to target `$tmp_parent_dir` instead of `$DIRECTORY`. The use of `$DIRECTORY` is not correct, as it is not set either by the test or the test system, and is empty at test runtime. Rather, `$tmp_parent_dir` should be checked for existence and, if it doesn't exist, a default initial permissions value is set for later checks.