[YDB#1034] Test MUPIP should accept case-insensitive region names
- Modify the mupip/mu_extend test to use a mixed-case region name
- Modify the mupip/mu_create test to use a mixed-case region name
- Modify the reorg/restrict_region test to use a mixed-case region name
- Modify the v70003/integ_dumpfh_order-gtmf134692 test to use a list of
  mixed-case region names

To keep this simple, each MUPIP operation test across the test suite
(that didn't already do this) was modified to test a region name in
mixed-case, following the pattern for MUPIP operations that already had
tests that covered that case. In every one of the sub-tests that were
modified, there were still test cases using full upper-case region
names, so no test coverage was lost.