Fix very rare v71000/rctldump_superseded-gtmf135385 subtest hang (test timing issue)
For example, see this `ps -ef --forest` output for a failing run of this test:

```
user      91061      1  0 Jul01 pts/0    00:00:00 tcsh -f -e YDBTESTPATH/v71000/instream.csh
user      91072  91061  0 Jul01 pts/0    00:00:00  \_ /usr/local/bin/tcsh -f YDBTESTPATH/com/submit_subtest.csh
user      11198  91072  0 Jul01 pts/0    00:00:00      \_ tcsh -f YDBTESTPATH/v71000/u_inref/rctldump_superseded-gtmf135385.csh
user      11399  11198  0 Jul01 pts/0    00:00:01          \_ YDB_DIST/mumps -run runTest^gtmf135385
user      11315      1  0 Jul01 pts/0    00:00:01 YDB_DIST/mumps -run runTest^gtmf135385
```

In the above case, process `11315` was expected to hang by the test case, but `11399` was not. Rather, `11399` should have `write`n a number and exited immediately.

Since process `11399` is not expected to be hung by the test case, it is also not terminated by the test case, causing it to hang indefinitely.

We have a theory on the cause of this longstanding test timing issue: namely, that it is caused by the compilation of `gtmf135385.m` on line `38` completing after overwriting the routine on line `54`, but before compiling it on line `55`. As result, line `55` instead loads `gtmf135385.o` from the compilation of the routine on line `36` instead of the routine generated on line `54`.

```csh
35 echo '# Generate the first test routine version in src/gtmf135385.m'
36 echo 'runTest write $zroutines set $ZINTERRUPT="zhalt 1" for i=1:1  hang 1' > src/gtmf135385.m
37 echo '# Run the first version of the test routine in the background and capture the PID in gtmf135385-v1.pid'
38 ($gtm_exe/mumps -run runTest^gtmf135385 & ; echo $! >&! gtmf135385-v1.pid) >&! gtmf135385-v1.out
...
53 foreach i (`seq 1 5`)
54         echo "runTest write $i" >& src/gtmf135385.m
55         $gtm_exe/mumps -run runTest^gtmf135385 >&! gtmf135385-v$i.out
56         $gtm_exe/mupip rctldump >>& rctldump.out
57 end
```

The fix is to precompile the initial version of the `gtmf135385.m` routine before running it on line `38` to ensure that compilation completes before generating any new versions of the routine on line `54`.

Also, this test revises the routine version numbers to prevent test output from being overwritten for version `1`, since the previous version of the test used this version number for both the first and second versions of the `gtmf135385.m` routine.