[Pipeline] Check pipeline-test-ydbtest.csh against all YDB MRs not just open ones
This commit is a follow-up to commit
efe7819f178021918825f7dadada3224c2418f4f.

We found out that in our workflow, we frequently merge the YDB MR PRIOR
to creating the YDBTest MR; and thus the pipeline would fail because it
can't find the YDB MR. This commit changes the behavior of the pipeline
to look for ALL recent YDB MRs, rather than only those that are open.

Brief testing shows that the API returns by default the last 20 MRs in
reverse chronological order. The last MR created shows up first.

Note that a successful pipeline still requires that the author does not
delete their branch.