[#703] [V7.1-003] [GTM-DE563049] Enhance tls/gtmtlslib_cipherversion-gtmf167995 to test prevention of inappropriate modification of TLS protocol cipher options
This commit revises `tls/inref/gtmf167995.m` to omit previous logic to include TLS v1.3 cipher options in the outref check when a TLS version of 1.2 was passed to `getciphers()`. This was previously done to workaround the TLS minimum and maximum protocol versions that were not reset by `getciphers()`. Since this reset is done with the v7.1-003 changes, it is no longer necessary to include the TLS 1.3 cipher options when checking TLS 1.2 cipher option output.

Also, the outref was updated to accomodate multiple versions of `libgtmtls.so`.

Finally, SUSE-only failures due to the `lsb_command` being missing by default on SUSE systems was fixed by revising a `ZSYSTEM` command to use `cat /etc/os-release` instead.