[#852][YDB#1162] Update r134/ydb772 due to change in behavior
This was missed in commit 5e196338eaaf33c358560fbff4155101cda1c7d8 and
was found in subsequent D_ALLs.

We now follow symbolic links in %RD/%RSEL. Uses of `SOURCE_PATH` in this
test should now be `SOURCE_REALPATH`. Some usages of `SOURCE_PATH` come
from the shell script and are not resolved to realpath, so these should
just stay the same.