[#787] Fix the encryption failure on pipeline
Per this
https://gitlab.com/YottaDB/DB/YDBTest/-/issues/787#note_2701216171, the
Encryption tests fail because when we change the paths to the testareas
in the pipeline, the socket name ends up being too long. The important
error message is:

```
$ gpg-agent -vvv --debug-all --daemon --batch
...
gpg-agent[1151]: socket name '/builds/shabiel/YDBTest/testarea1/tst_V999_R999_dbg_01_250821_144607_n22/encryption_0/tmp/.gnupg/S.gpg-agent' is too long
```

This error message was found after a lot of debugging and the insertion
of the above `gpg-agent` command. The actual error message that shows up in
the test output looks like this:

```
$ cat -n testarea1/tst_V999_R999_dbg_01_250811_172728_xxL/encryption_1/err_mu_replic/create_key_file.out
    119 $gtm_tst_ver_gtmcrypt_dir/gen_sym_key.sh 0 $key
    120 /usr/library/V999_R999/dbg/plugin/gtmcrypt/gen_sym_key.sh 0 mumps_dat_key
    121 gpg: error running '/usr/bin/gpg-agent': exit status 2
    122 gpg: failed to start agent '/usr/bin/gpg-agent': General error
    123 gpg: can't connect to the agent: General error
    124 gpg: keydb_search failed: No agent running
    125 gpg: cannot open '/dev/tty': No such device or address
```

Line 125 is misleading. GPG is just badly designed, and apparently won't
print an error message unless you give it a command line `--batch`. If
you do, the error message you see will vary based on the GPG version.
The current version (gpg (GnuPG) 2.4.7) says something to the effect
that a user cannot be found, which is a bit non-sensical. `--batch` was
added to the main file (`com/create_key_file.csh`) that runs GPG, but
not everywhere.

This MR fixes this by configuring GPG to use specific Unix Socket paths
rather than ones located in the test directory. This is done in
`docker/setup_gpg.csh`.

While this fixes most tests, there was a simple problem with
`header_verify` and a much harder to find problem that ended up being in
`sym_key_management`.

* `header_verify` greped for DB to look for specific DSE output. The
  gitlab pipeline by definition uses the project name in the directory
  (YDBTest), and DB is part of YDBTest. That resulted in printing out
  DSE output that was not intened. This is now fixed by a `grep
  -v`
* `sym_key_management` is one of the rare tests that modifies the gpg
  directory on the fly. It did that without stopping the gpg-agent that
  is currently running in the background. Apparently this issue was
  innocuous when there were no socket configurtaions for GPG, but now
  that there is, gpg starts getting confused and exhibits symptoms of it
  trying to use the full path for sockets. The solution is simple: kill
  the gpg-agent before starting to modify the gpg directory.

Misc changes:

* In the rsync done to move test system files to $gtm_tst, exclude the
  .git folder. It's not needed and slows us down.
* A small whitespace change suggested by the pre-commit script.

Note that the pipeline on this commit will fail as it requires a
re-building of the docker image. The fix was verified using a local
build of the pipeline image and the
https://github.com/firecow/gitlab-ci-local tool.