[#787] Fix 4 tests on Docker
Tests:

- v60001/gtm7510
- v60001/gtm7552
- v62001/gtm8106
- v62001/gtm8191a

See https://gitlab.com/nars1/YDB/-/jobs/11322355319 for an example
failure.

The cause is that the test assumes dse/mupip/lke will be on $PATH or an
alias.  That's not a valid assumption. The fix is is to use the full
path by pre-pending $gtm_dist.

Misc: White space changes to change spaces to tabs per our conventions.