[YDB#1130]  New r204/valtoobig-ydb1130 Test for VALTOOBIG error
This test exist because there was an error with displaying the VALTOOBIG error
in the context of setting allocation and autoswitch to produce one.

A typo in the quotes in one line ensured that the function called to output the error would encounter its own new error instead of outputting the correct error.

This test will activate that code and ensure that the correct error, a val too big error, will occur.