[YDB#1170] Fix use of --linkexec in sudo/tmpyottadbperms-ydb1125
After merging of YDB#1170, we started getting test failures from the
above test containing the following diff line:

```
> ls: cannot access '/yottadb': No such file or directory
```

Investigation revealed that this is due to an erroneous use of
`--linkexec dirname`. The linkexec directory was the same as the install
directory, resulting in a symbolic link for `yottadb` that overwrote
itself and referenced itself.

Why does the test use a symbolic link if it doesn't use it? This message
says why:
```
Use --linkexec to reduce instance of failures in concurrently running tests that access /usr/local/bin
```

`ydbinstall` creates symlinks by default in /usr/local/bin if no other
options are specified. So using `--linkexec dir` will create the links
elsewhere, letting us run concurrent tests.

The solution is to have a linkexec directory that is a throwaway
directory.