[Pipeline] Improvement to open MR checking for pipeline-test-ydbtest.csh
Commit 7240bada269595f42a91746a9bfe36a2d48a61d0 added logic to get all
MRs in YDB to compare against YDBTest by branch name. This worked well
in almost all cirumstances. However, we found that if you work on a YDB
MR for a long time--as the API only gets the last 20 MRs--your still
open YDB MR may page off the list.

See
https://gitlab.com/YottaDB/DB/YDBTest/-/merge_requests/2378#note_2827102467
for a background.

This MR changes the logic as follows:

- See if there is an open YDB MR with the source_branch branch name. If
  so, use it.
- If not, search for last 20 merged YDB MRs and see if desired branch
  name is found. If so, use it.
- If not, use YDB master.

Note that unlike the old logic, we don't get closed MRs. That's okay, as
they are rare and not practically useful (we don't want to test against
a close YDB MR).