[YDB#933] New r204/job_cmdline_dash-ydb933 to test that the command line of a JOBed process can contain arguments that start with a dash.
Background
----------
There was a bug where a process started with a JOB command would have its command line evaluated the same as any other process.
This meant that if the command line started with an argument that started with a dash, the program would report a command not found error and never run.

Expected behavior
----------------
After the fix, the command line of a JOBed command will be ignored when parsing it for relevant commands.
You can retrieve this command line with $zcmdline.

Test
----
There are 2 parts to the test.
    - First, it tests a JOBed command with a command line that starts with a dash to ensure that the JOB command runs.
    - Second, it runs YDB -dir -port as this should still give a YDB-E-CLIERR error.
            -For more on why this is still an error for now, see https://gitlab.com/YottaDB/DB/YDB/-/merge_requests/1792#note_2951733798