[#874] Fix rare r204/naked_refs-ydb665 subtest failure (INTRPT issued line out of order)
Background
----------

Previously, the `r204/naked_refs-ydb665` subtest failed with the following diff:

```diff
581d580
< INTRPT issued to process NNNN
583a583
> INTRPT issued to process NNNN
```

This was due to a timing issue that resulted in the INTRPT output appearing in a different sequence in some test runs. Specifically, the test sends a MUPIP INTRPT to a backgrounded MUMPS process that issues output after being interrupted.

However, it was possible for the interrupt to be received and the backgrounded MUMPS process to issue its output before the output of the MUPIP INTRPT process running in the foreground made it to the terminal, thus resulting in output being out of order in some cases.

Fix
---

This commit fixes this issue by logging the output of the backgrounded MUMPS to a file and only printing that to the terminal (via `cat`) after the MUPIP INTRPT has been issued and the backgrounded MUMPS process dies. This ensures that the output is always in the same order in every test run.

This change was accomplished by refactoring the affected portion of the test to use `tcsh` backgrounding conventions and `wait_for_proc_to_die.csh` as is done in other tests.

Also, multiple M routines embedded in `echo` statements were moved into a single `inref/ydb665.m` fixture for readability, and additional test flow comments were added to further clarify the test case in question.

Test
----

Prior to this change, the above failure would occur in about 16/20 test runs on one in-house system. Now, this failure is seen in 0/20 test runs.