[YDB#858] New r204/generated_from_msg-ydb858 to test syslog message shows library base + offset
Accompanying commit summary
---------------------------
The YDB commit #858 improves the message generated by send_msg_va() to include the library base and offset address.
It should also include the name of the function if it is not local, but most function names will not be accessible.

r204/generated_from_msg-ydb858
------------------------------
The test creates a database and tests freezing and unfreezing that database
as this action should write the message being tested to the system logs.
As the message includes the test path, the test runs grep $tst_working_dir to
ensure that it does not pick up any background messages in the system logs from other processes.
Then, it saves the relative addresses and uses the mumps program ydb858.m
to launch the gdb debugger and check the exact positions in the code indicated by the relative address
to ensure it finds the line that created the "generated from" messages.

v63007/gtm7318
--------------
As stated in the Accompanying commit summary, if the function name is
available to the code, it prints the function_name+offset instead.
In order to check this functionality, v63007/gtm7318 has been modified to check the syslogs
as it was already printing this message to the syslogs.
This will be how we test printing the function name instead of doing it in r204/generated_from_msg-ydb858.

v60000/gtm7327
--------------
This test was expecting many "generated from" messages. It now expects them to include the new base+offset format.

stress/concurr_replwason
------------------------
This test was expecting one "generated from" message. It now expects it to include the new base+offset format.

online_rollback/basic
---------------------
This test was expecting many "generated from" messages.
One of them generates a function that is exported and so it expects the new symbolic_name+offset format.
The rest of the "generated from" messages expect the more common base+offset format.

YDB MR link
-----------
https://gitlab.com/YottaDB/DB/YDB/-/merge_requests/1799