[YDB#1201] Fix sudo/ydb910 subtest failure (no space on /tmp for YDB build from source)
This fixes a failure on small systems (all ARM-based) that don't have
enough space on `/tmp` for `ydbinstall.sh` to compile YottaDB.

Minor changes to the `.csh` test file as the install artifacts still did
not zip sucessfully due to the lack of write permissions on the folders.

Whitespace changes requested by the pre-commit hook added.