[YDB#917] New r204/mupip_terminal_prompts-ydb917 to test various mupip commands with empty -file argument
Overview
------------
Previously, if you called any of 5 different mupip commands with an empty -file argument, the prompt "File or Region" would appear.
However, only a filename would work.
After YDB#917 it now only says to enter a file.
The 5 mupip commands are dumpfhead, integ, upgrade, set, and rundown.
Mupip dumpfhead and mupip integ would exhibit the same behavior if instead of an empty -file argument,
they were invoked without any arguments.
Additionally, only in the mupip dumpfhead case, there was a bug
resulting in
it not printing the newline after the prompt until after printing the output.
This bug has also been fixed in YDB#917.
All of this is tested in r204/mupip_terminal_prompts-ydb917.
Additionally, some tests were expecting the old message, they have been fixed to expect the new one.

Test details
---------------
The test uses heredocs to test inputting a fake filename and real region name into all 5 mupip utilities.
It makes sure that all the fake filenames show the correct prompt, not asking for a region name.
It makes sure that all the real region names produce an error, ensuring that a region is indeed an invalid input for these prompts.
It runs the same test for mupip dumpfhead and mupip integ without the -file argument.
By testing the mupip dumpfhead prompt, we automatically check that the newline placement bug is fixed in the basic case.
To make sure that it is fixed in every case, 2 expect scripts are used to test mupip dumpfhead with a terminal input.

r132/ydb635
------------
This test was bringing up the old message "File or Region:".
It now expects just the message "File:".
Additionally, the echo statements in this file have been updated to reflect this.

v62002/gtm8228
--------------
This test was bringing up the old error message "File or Region:".
It now expect just the message "File:"

Gitlab MR link
--------------
https://gitlab.com/YottaDB/DB/YDB/-/merge_requests/1803