[#720] Fix history corruption due to gtmtest alias
With the use of the developer-friendly environment changes introduced
in [#720], we had a very difficult to identify issue having to deal with
history being re-written on the fly.

The saved history showed entries like this when you issued this command:
`gtmtest -t basic`:

- `-t basic#+1767198460`

The runtime history (accessed by the up-arrow and Ctrl-R) showed this:

- `-t`

The problem was due to the gtmtest alias:

```
alias gtmtest 'tsync; tcsh -fc "$gtm_tst/com/gtmtest.csh $gtmtest_args \!*:q"; source $gtm_test_com_individual/lasttest.csh'
```

`!*:q` took the `-t basic` and added it to the command line, and then as
well to the history.

There are multiple possible fixes. A fix suggested by AI is to put the
output of `!*` in a variable, which does not get added to history. A
better fix (not pursued) is to make the alias `gtmtest` call a script.